#2227 ThemeAPI - rationalizing duplicated methods between checkout and customer classes.
This commit should mean all methods that were previously duplicated are now housed only in ShoppCustomerThemeAPI. There are still some others where it might make sense to move them across though that perhaps should be part of a broader sweep.

Todo: set of testing to ensure A) most critically that checkout does not break in any way as a result of these changes B) in the best case work to make a custom registration form viable again hasn't been compromised.