Fixed DownloadAsset issue when multiple wp_shopp_meta records exist
When attaching a new downloadable file to a product, a new
wp_shopp_meta line is generated for that product. The SQL should be
returning the *latest* file (so Purchases from before the change still
work) but it's not. Fixed it by adding an ORDER BY clause (and also
changing to an INNER JOIN).