Suppress purposeful use of array_diff for multi-dimensional array input cases
Noted that array_diff can cause string to array conversion PHP Notice
when a shipping rule for a country has multiple country areas defined.
This is intentional to get the fast performance of array_diff to
determine differences in the rule and the potential match target.
Country area matching is done later to resolve those differences.