Restored the default notice.
This is intended. The error is not this low-level call, but rather that the Advanced screen and other screen controllers that extend this class have not been cleaned up to follow the new paradigm. They were committed in a state of minimum viable operations and now need finalized in order to override the low-level notice.