Changed approach for detecting realtime shipping service providers
We detect realtime shipping service providers to determine whether to offer separate packaging for a product. This information can be requested from the core Shopp object. The new screen metabox system means we can add the reference at the metabox controller level and provide a simple data reference variable for the UI markup.